Move tests to subdir & add model tests.
