+++ tests/notifications/tests-notification-model.php
@@ -1,0 +1,58 @@
+<?php
+/**
+ * Notification Model Tests
+ *
+ * @package   easy-digital-downloads
+ * @copyright Copyright (c) 2021, Easy Digital Downloads
+ * @license   GPL2+
+ */
+
+namespace EDD\Tests\Notifications;
+
+use EDD\Models\Notification;
+
+/**
+ * @coversDefaultClass \EDD\Models\Notification
+ * @group edd_notifications
+ */
+class NotificationModelTests extends \EDD_UnitTestCase {
+
+	private function insertAndGetNotification( $data ) {
+		$notificationId = EDD()->notifications->insert( $data, 'notification' );
+
+		return new Notification( EDD()->notifications->get( $notificationId ) );
+	}
+
+	/**
+	 * The `dismissed` property should be cast to boolean.
+	 *
+	 * @covers \EDD\Models\Notification::castAttribute
+	 */
+	public function test_dismissed_is_boolean() {
+		$notification = $this->insertAndGetNotification( array(
+			'title'     => 'Notification',
+			'content'   => 'Notification',
+			'type'      => 'success',
+			'dismissed' => 0,
+		) );
+
+		$this->assertFalse( $notification->dismissed );
+	}
+
+	/**
+	 * The `id` property should be cast to an integer.
+	 *
+	 * @covers \EDD\Models\Notification::castAttribute
+	 */
+	public function test_id_is_integer() {
+		$notification = $this->insertAndGetNotification( array(
+			'title'     => 'Notification',
+			'content'   => 'Notification',
+			'type'      => 'success',
+			'dismissed' => 0,
+		) );
+
+		$this->assertTrue( is_int( $notification->id ) );
+	}
+
+}
